os: fix hex exit code print on 32-bit windows
We want to print hex exit codes for the large values,
but on 32-bit Windows the large values are negative.

Reviewed-on: https://go-review.googlesource.com/c/go/+/294070
Trust: Russ Cox <rsc@golang.org>

+++ src/os/exec_posix.go
@@ -103,7 +103,7 @@ func (p *ProcessState) String() string {
 	switch {
 	case status.Exited():
 		code := status.ExitStatus()
-		if runtime.GOOS == "windows" && code >= 1<<16 { // windows uses large hex numbers
+		if runtime.GOOS == "windows" && uint(code) >= 1<<16 { // windows uses large hex numbers
 			res = "exit status " + uitox(uint(code))
 		} else { // unix systems use small decimal integers
 			res = "exit status " + itoa(code) // unix
